Balanced Fatigue better Fatigue still requires manual input from words Refactored Entity's IEnergyProducer wrapper to work with Entity, EntityLivingBase, and EntityPlayer Fixed up first IModifier not working with reasoning Energy Requests now work on float percentages from 0.0f to 1.0f

+++ src/main/java/com/firegodjr/ancientlanguage/api/magic/IEnergyProducer.java
@@ -4,14 +4,15 @@ package com.firegodjr.ancientlanguage.api.magic;
  * An interface for objects implementing magical energy production
  */
 public interface IEnergyProducer {
-	
+
 	/**
 	 * Handles magic use
-	 * 
+	 *
 	 * @param energyToPull
-	 *            The amount of energy to pull
-	 * @return The amount of energy retrieved from energy request
+	 *            The percentage of energy to pull
+	 *
+	 * @return The percentage of asked energy retrieved from energy request
 	 */
-	public int useMagic(int energyToPull);
-	
+	public float useMagic(float energyToPull);
+
 }
